Refactor: Encapsulate logic for patient endpoint
