+++ broker/server.js
@@ -3,8 +3,7 @@ const configs = require("./config/server.js");
 
 const express = require("express");
 const axios = require("axios").default;
-const Patient = require("./models/Patient");
-const RelatedPerson = require("./models/RelatedPerson");
+const Patient = require("./endpoints/Patient");
 const Organization = require("./models/Organization");
 const Encounter = require("./models/Encounter");
 const Observation = require("./models/Observation");
@@ -28,7 +27,6 @@ if (process.env.DEVELOPMENT == 1) {
   });
 }
 
-const base = "http://hapi:8080/hapi-fhir-jpaserver/fhir";
 const generalEndpoints = [
   "/Encounter*",
   "/Observation*",
@@ -36,9 +34,6 @@ const generalEndpoints = [
   "/Appointment*",
   "/Immunization*",
 ];
-const headers = {
-  "content-type": "application/fhir+json",
-};
 const patchHeaders = {
   "content-type": "application/json-patch+json",
 };
@@ -51,7 +46,7 @@ app.get("/healthcheck", (req, res) => {
 // Pass GET requests to HAPI FHIR server
 app.get(generalEndpoints, (req, res) => {
   axios
-    .get(`${base}${req.url}`)
+    .get(`${configs.base}${req.url}`)
     .then((response) => {
       res.json(response.data);
     })
@@ -62,24 +57,8 @@ app.get(generalEndpoints, (req, res) => {
     });
 });
 
-app.get("/Patient*", (req, res) => {
-  axios
-    .get(`${base}${req.url}`)
-    .then((response) => {
-      let r;
-      if (response.data.resourceType === "Bundle") {
-        r = response.data.entry.map((entry) => Patient.toModel(entry.resource));
-      } else {
-        r = Patient.toModel(response.data);
-      }
-      res.json(r);
-    })
-    .catch((e) => {
-      res.status(400).json({
-        error: e.response ? e.response.data : e.message,
-      });
-    });
-});
+app.post("/Patient", Patient.create);
+app.get("/Patient*", Patient.read);
 
 // Create Immunization resource
 // Response:  Immunization resource (200)
@@ -100,9 +79,11 @@ app.post("/Immunization", (req, res) => {
 async function postImmunization(imm) {
   let resource = Immunization.toFHIR(imm);
 
-  return axios.post(`${base}/Immunization`, resource).then((response) => {
-    return response;
-  });
+  return axios
+    .post(`${configs.base}/Immunization`, resource)
+    .then((response) => {
+      return response;
+    });
 }
 
 app.post("/Appointment", (req, res) => {
@@ -110,7 +91,7 @@ app.post("/Appointment", (req, res) => {
   let resource = Appointment.toFHIR(appt);
 
   axios
-    .post(`${base}/Appointment`, resource)
+    .post(`${configs.base}/Appointment`, resource)
     .then((response) => {
       res.json(response.data);
     })
@@ -126,7 +107,7 @@ app.post("/Organization", (req, res) => {
   let resource = Organization.toFHIR(org);
 
   axios
-    .post(`${base}/Organization`, resource)
+    .post(`${configs.base}/Organization`, resource)
     .then((response) => {
       res.json(response.data);
     })
@@ -157,25 +138,26 @@ app.post("/Observation", (req, res) => {
 async function postObservation(observation) {
   let resource = Observation.toFHIR(observation);
 
-  return axios.post(`${base}/Observation`, resource).then((response) => {
-    let ref = observation.partOf;
-    // request body for PATCH
-    let update = [
-      {
-        op: "add",
-        path: "/reaction",
-        value: [{ detail: { reference: `Observation/${response.data.id}` } }],
-      },
-    ];
-    let headers = {
-      "content-type": "application/json-patch+json",
-    };
+  return axios
+    .post(`${configs.base}/Observation`, resource)
+    .then((response) => {
+      let ref = observation.partOf;
+      // request body for PATCH
+      let update = [
+        {
+          op: "add",
+          path: "/reaction",
+          value: [{ detail: { reference: `Observation/${response.data.id}` } }],
+        },
+      ];
 
-    axios.patch(`${base}/${ref}`, update, { headers: headers }).catch((e) => {
-      console.log({ error: e.response ? e.response.data : e.message });
+      axios
+        .patch(`${configs.base}/${ref}`, update, { headers: patchHeaders })
+        .catch((e) => {
+          console.log({ error: e.response ? e.response.data : e.message });
+        });
+      return response;
     });
-    return response;
-  });
 }
 
 app.post("/Encounter", (req, res) => {
@@ -183,7 +165,7 @@ app.post("/Encounter", (req, res) => {
   let resource = Encounter.toFHIR(encounter);
 
   axios
-    .post(`${base}/Encounter`, resource)
+    .post(`${configs.base}/Encounter`, resource)
     .then((response) => {
       res.json(response.data);
     })
@@ -193,73 +175,6 @@ app.post("/Encounter", (req, res) => {
       });
     });
 });
-
-app.post("/Patient", (req, res) => {
-  createPatient(req, res).catch((e) =>
-    res.status(400).json({
-      error: e.response ? e.response.data : e.message,
-    })
-  );
-});
-
-async function createPatient(req, res) {
-  let patients = req.body.Patient;
-  let patient = patients.shift();
-  let promises = [];
-  let response = { Patient: [] };
-
-  // resource for head of household
-  if (patient) {
-    let related = { resourceType: "RelatedPerson" };
-    let result = await createPatientWithLink(patient, related);
-    response.Patient.push(result.data);
-
-    // update related
-    let related_id = result.data.link[0].other.reference.split("/")[1];
-    resource = RelatedPerson.toFHIR({
-      patient: `Patient/${result.data.id}`,
-      relationship: patient.relationship,
-    });
-    resource.id = related_id;
-    await axios.put(`${base}/RelatedPerson/${related_id}`, resource, headers);
-  } else {
-    res.status(400).json({ error: "at least one patient should be provided" });
-    return;
-  }
-
-  // resources for members
-  for (patient of patients) {
-    let related = {
-      patient: `Patient/${response.Patient[0].id}`,
-      relationship: patient.relationship,
-    };
-
-    promises.push(createPatientWithLink(patient, related));
-  }
-
-  Promise.all(promises).then((results) => {
-    for (result of results) {
-      response.Patient.push(result.data);
-    }
-    res.json(response);
-  });
-}
-
-// create patient and link to relatedperson
-async function createPatientWithLink(patient, related) {
-  // create related person
-  let resource = RelatedPerson.toFHIR(related);
-  let related_id = (
-    await axios.post(`${base}/RelatedPerson`, resource, headers)
-  ).data.id;
-
-  // create patient pointing to related person
-  patient.link = [`RelatedPerson/${related_id}`];
-  resource = Patient.toFHIR(patient);
-  let result = await axios.post(`${base}/Patient`, resource, headers);
-
-  return result;
-}
 
 // Discharge
 app.post("/discharge", (req, res) => {
@@ -333,7 +248,7 @@ app.post("/check-in", (req, res) => {
 // URL is either a specific id '/id'
 // or query parameters '?param=value'
 function updateEncounterStatus(url, status) {
-  return axios.get(`${base}/Encounter${url}`).then((response) => {
+  return axios.get(`${configs.base}/Encounter${url}`).then((response) => {
     let encounter;
     let resourceType = response.data.resourceType;
     let patch;
@@ -359,7 +274,7 @@ function updateEncounterStatus(url, status) {
 
     // update the database with new encounter
     return axios
-      .patch(`${base}/Encounter/${encounter.id}`, patch, {
+      .patch(`${configs.base}/Encounter/${encounter.id}`, patch, {
         headers: patchHeaders,
       })
       .then((response) => {
@@ -369,7 +284,7 @@ function updateEncounterStatus(url, status) {
 }
 
 function updateAppointmentStatus(url, status) {
-  return axios.get(`${base}/Appointment${url}`).then((response) => {
+  return axios.get(`${configs.base}/Appointment${url}`).then((response) => {
     let appt;
     let resourceType = response.data.resourceType;
     let patch;
@@ -395,7 +310,9 @@ function updateAppointmentStatus(url, status) {
 
     // update the database with new appointment
     return axios
-      .patch(`${base}/Appointment/${appt.id}`, patch, { headers: patchHeaders })
+      .patch(`${configs.base}/Appointment/${appt.id}`, patch, {
+        headers: patchHeaders,
+      })
       .then((response) => {
         return response;
       });
