--- models/CreateNewEncounterRequest.js
@@ -1,28 +1,0 @@
-/*
-Encounter {
-  status : enum (planned, arrived, in-progress, finished, cancelled)
-  class : enum (FLD)
-  subject : string
-  appointment : string
-  period : {
-    start : datetime
-    end : datetime
-  }
-  location : string
-}
-{
-  "Encounter": [
-    {
-      "status" : "planned"
-      "class" : "FLD"
-      "subject" : "1465563"
-      "appointment" : "1703345"
-      "period" : {
-        "start" : "2020-09-17T05:10:19+00:00"
-        "end" : "2020-09-17T05:15:19+00:00"
-      }
-      location : "1474953"
-     }
-   ]
-}
-*/
