--- models/CreateNewEpisodeOfCareRequest.js
@@ -1,16 +1,0 @@
-/*
-EpisodeOfCare {
-  status : enum (planned, active, finished)
-  patient : string
-  managingOrganization : string
-}
-{
-  "EpisodeOfCare": [
-    {
-      "status" : "planned"
-      "patient" : "1457642"
-      "managingOrganization" : "1703654"
-     }
-   ]
-}
-*/
