--- models/CreateNewManagingOrganizationRequest.js
@@ -1,16 +1,0 @@
-/*
-Organization {
-  active : boolean
-  type : enum (govt, edu, reli, cg, other)
-  name : string
-}
-{
-  "Organization": [
-    {
-      "active" : "true"
-      "type" : "cg"
-      "name" : "ManagingOrganizationName"
-     }
-   ]
-}
-*/
