--- models/CreateNewManufacturerOrganizationRequest.js
@@ -1,16 +1,0 @@
-/*
-Organization {
-  active : boolean
-  type : enum (prov, bus, other)
-  name : string
-}
-{
-  "Organization": [
-    {
-      "active" : "true"
-      "type" : "other"
-      "name" : "ManufacturerOrganizationName"
-     }
-   ]
-}
-*/
