--- models/CreateNewOrganizationRequest.js
@@ -1,15 +1,0 @@
-/*
-Organization {
-  active : boolean
-  type : enum (prov, bus, govt, edu, reli, cg, other)
-  name : string
-}
-{
-  "Organization":
-    {
-      "active" : true,
-      "type" : "cg",
-      "name" : "ManagingOrganizationName"
-    }
-}
-*/
